requirements.txt
py-cord==2.0.0b7
python-decouple~=3.5
requests~=2.27.1
pymongo~=3.12.3
pymodm~=0.4.3